Add agent options.
